+++ main.py
@@ -75,6 +75,8 @@ if config:
     max_height = config['IMAGE_SPLIT']['max_height']
     # For translation
     target_language = config['TRANSLATION']['target_language']
+    max_retries = config['TRANSLATION']['max_retries']
+    retry_delay = config['TRANSLATION']['delay']
     gemini_model = config['TRANSLATION']['gemini']['model']
     gemini_temp = config['TRANSLATION']['gemini']['temperature']
     gemini_top_p = config['TRANSLATION']['gemini']['top_p']
@@ -271,7 +273,23 @@ for dirpath, dirnames, filenames in natsorted(os.walk(args.input)):
             recognitions.extend(recognition)
 
     # --- Stage 5/3: Translate Extracted Text with Gemini ---
-    translated_text_data = translate_texts_with_gemini(recognitions, target_language, [gemini_model, gemini_temp, gemini_top_p, gemini_max_out_tokens], previous_dir, output_dir, log_level)
+    # Use automatic retry in case any errors with translation
+    max_retries = max_retries
+    retry_delay = retry_delay
+    attempts = 0
+
+    while attempts < max_retries:
+        try:
+            translated_text_data = translate_texts_with_gemini(recognitions, target_language, [gemini_model, gemini_temp, gemini_top_p, gemini_max_out_tokens], previous_dir, output_dir, log_level)
+            break
+        except Exception as e:
+            attempts += 1
+            logger.error(Fore.RED + f"{e}")
+            if attempts < max_retries:
+                logger.info(f"({attempts}/{max_retries}) Retrying in {retry_delay} seconds...")
+                time.sleep(retry_delay)
+            else:
+                raise Exception(Fore.RED + "Max retries reached!")
 
     # --- Stage 6/4: Whiten Text Areas & Overlay Translated Texts to Split Images ---
     overlay_translated_texts(image_chunks, merge_images, translated_text_data, [box_offset, box_padding, box_fill_color, box_outline_color], [font_min, font_max, font_color, font_path], common_original_extension, [source_language, lang_code_jp], output_dir, log_level)